+++ src/Loader/Dependency.js
@@ -9,7 +9,7 @@ const Options = require('./Options');
 class Dependency {
   /** The file system used by Webpack */
   static fileSystem = null;
-  static files = null;
+  static files = new Set();
   static loaderContext = null;
   static watchFiles = {};
   static #entryFiles = [];
@@ -17,20 +17,23 @@ class Dependency {
   static init(loaderContext) {
     if (!PluginService.isWatchMode()) return;
 
+    PluginService.setDependencyInstance(this);
     this.loaderContext = loaderContext;
     this.fileSystem = loaderContext.fs.fileSystem;
     this.watchFiles = Options.getWatchFiles();
     this.#entryFiles = AssetEntry.getEntryFiles();
     this.addFile = this.addFile.bind(this);
-    this.files = new Set();
 
     const fs = this.fileSystem;
-    const { files: includes, ignore: excludes } = this.watchFiles;
+    const { files: includes, ignore: excludes, paths = [] } = this.watchFiles;
 
-    for (const watchDir of this.watchFiles.paths) {
+    for (const watchDir of paths) {
       const files = readDirRecursiveSync(watchDir, { fs, includes, excludes });
       files.forEach(this.addFile);
     }
+
+    const customWatchFiles = Options.getCustomWatchFiles();
+    if (customWatchFiles.length > 0) customWatchFiles.forEach(this.addFile);
   }
 
   /**
@@ -62,6 +65,13 @@ class Dependency {
         loaderContext.addDependency(file);
       }
     }
+  }
+
+  /**
+   * Called when the compiler is closing or a watching compilation has stopped.
+   */
+  static shutdown() {
+    return this.files.clear();
   }
 }