+++ test/cases/msg-exception-loader-data-file-invalid/webpack.config.js
@@ -1,0 +1,21 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: './src/home.html',
+      },
+      loaderOptions: {
+        data: './data-global.json',
+      },
+    }),
+  ],
+};