+++ test/cases/msg-exception-plugin-option-afterProcess/webpack.config.js
@@ -1,0 +1,37 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+    publicPath: '/',
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: './src/index.html',
+      },
+      js: {
+        filename: 'assets/js/[name].[contenthash:8].js',
+      },
+      css: {
+        filename: 'assets/css/[name].[contenthash:8].css',
+      },
+      afterProcess: (content, { sourceFile, assetFile }) => {
+        throw new Error('Custom after process failed!');
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader', 'sass-loader'],
+      },
+    ],
+  },
+};