+++ test/cases/option-entry-data-file/src/views/home/data.js
@@ -1,0 +1,6 @@
+// you can get here data from database
+const data = {
+  title: 'Home',
+};
+
+module.exports = data;